+++ packages/server/.eslintrc.json
@@ -1,17 +1,25 @@
 {
-  "env": {
-    "node": true,
-    "es2021": true,
-    "jest": true
-  },
-  "extends": ["eslint:recommended", "prettier"],
-  "plugins": ["prettier"],
-  "parserOptions": {
-    "ecmaVersion": "latest",
-    "sourceType": "module"
-  },
+  "root": true,
+  "parser": "@typescript-eslint/parser",
+  "plugins": ["@typescript-eslint", "prettier"],
+  "extends": [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/eslint-recommended",
+    "plugin:@typescript-eslint/recommended",
+    "prettier"
+  ],
   "rules": {
-    "prettier/prettier": "error",
+    "no-console": 1,
+    "prettier/prettier": 2,
     "eol-last": "error"
-  }
+  },
+  "overrides": [
+    {
+      "files": ["**/*.test.ts"],
+      "env": {
+        "jest": true,
+        "node": true
+      }
+    }
+  ]
 }
