+++ packages/server/babel.config.json
@@ -7,6 +7,7 @@
           "node": "current"
         }
       }
-    ]
+    ],
+    "@babel/preset-typescript"
   ]
 }
