+++ packages/server/package.json
@@ -3,29 +3,42 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./index.js",
-    "start:mocked": "MOCKED=true node ./index.js",
+    "start": "yarn build",
+    "start:mocked": "MOCKED=true yarn build",
+    "build": "nodemon 'index.ts' --exec 'ts-node' index.ts -e ts,graphql",
     "test": "jest ./",
     "lint": "eslint ./",
     "prettier:fix": "yarn prettier --write ."
   },
   "dependencies": {
+    "@graphql-tools/schema": "^8.5.0",
+    "@types/graphql": "^14.5.0",
+    "@types/node": "^18.0.0",
     "apollo-server": "^3.9.0",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
     "express": "~4.16.1",
     "graphql": "^16.5.0",
+    "graphql-import": "^1.0.2",
+    "graphql-import-node": "^0.0.5",
+    "graphql-tools": "^8.3.0",
     "http-errors": "~1.6.3",
     "morgan": "~1.9.1"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@types/jest": "^28.1.3",
+    "@typescript-eslint/eslint-plugin": "^5.30.0",
+    "@typescript-eslint/parser": "^5.30.0",
     "babel-jest": "^28.1.1",
     "eslint": "^8.17.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "jest": "^28.1.1",
-    "prettier": "^2.7.1"
-  },
-  "type": "module"
+    "nodemon": "^2.0.18",
+    "prettier": "^2.7.1",
+    "ts-node": "^10.8.1",
+    "typescript": "^4.7.4"
+  }
 }
