+++ packages/server/resolvers/user.test.ts
@@ -1,4 +1,4 @@
-import { getUser } from './user.js';
+import { getUser } from './user';
 
 describe('user resolver', () => {
   describe('getUser', () => {
